+++ contracts/getting_started_with_solidity/02_constructors/MyContract.sol
@@ -5,15 +5,15 @@
 pragma solidity ^0.8.0;
 
 contract MyContract {
-  uint public x;
-  uint public y;
-  address public owner;
-  uint public createdAt;
+    uint256 public x;
+    uint256 public y;
+    address public owner;
+    uint256 public createdAt;
 
-  constructor(uint _x, uint _y) {
-    x = _x;
-    y = _y;
-    owner = msg.sender;
-    createdAt = block.timestamp;
-  }
+    constructor(uint256 _x, uint256 _y) {
+        x = _x;
+        y = _y;
+        owner = msg.sender;
+        createdAt = block.timestamp;
+    }
 }
