+++ contracts/getting_started_with_solidity/03_variables/SampleContract.sol
@@ -5,27 +5,27 @@
 pragma solidity ^0.8.0;
 
 contract SampleContract {
-  //state variable of data type unint declared called "storedData"
-  //Data in this variable will be saved to the block chain
-  uint public storedData;
+    //state variable of data type unint declared called "storedData"
+    //Data in this variable will be saved to the block chain
+    uint256 public storedData;
 
-  function getResult() public pure returns(uint) {
-    //local variable
-    uint a = 7;
-    //local variable
-    uint b = 3;
-    uint result = a + b;
-    //access the local variable
-    return result;
-  }
+    function getResult() public pure returns (uint256) {
+        //local variable
+        uint256 a = 7;
+        //local variable
+        uint256 b = 3;
+        uint256 result = a + b;
+        //access the local variable
+        return result;
+    }
 
-  function getSenderOfTransaction() public view returns(address) {
-    //access global variables
-    return msg.sender;
-  }
+    function getSenderOfTransaction() public view returns (address) {
+        //access global variables
+        return msg.sender;
+    }
 
-  function getChainId() public view returns(uint) {
-    //access global variables
-    return block.chainid;
-  }
+    function getChainId() public view returns (uint256) {
+        //access global variables
+        return block.chainid;
+    }
 }
