+++ contracts/getting_started_with_solidity/13_array/MyContract.sol
@@ -5,36 +5,36 @@
 pragma solidity ^0.8.0;
 
 contract MyContract {
-  uint[] public myArray; //this is a dynamic array of type uint
-  uint[] public myArray2 = [1, 2, 3, 4]; //this is a dynamic array with 1, 2 and 3 as default values
-  uint[10] public myFixedSizeArray; //this is a fixed size array of type uint
+    uint256[] public myArray; //this is a dynamic array of type uint
+    uint256[] public myArray2 = [1, 2, 3, 4]; //this is a dynamic array with 1, 2 and 3 as default values
+    uint256[10] public myFixedSizeArray; //this is a fixed size array of type uint
 
-  uint[] intergerArray;       //sample array of integers
-  bool[] boolArray;           //sample array of booleans
-  address[] addressArray;     //sample array of address
+    uint256[] intergerArray; //sample array of integers
+    bool[] boolArray; //sample array of booleans
+    address[] addressArray; //sample array of address
 
-  function pushIsToAdd(uint i) public {
-    myArray.push(i);
-  }
+    function pushIsToAdd(uint256 i) public {
+        myArray.push(i);
+    }
 
-  function getItemInArray(uint index) public view returns(uint) {
-    return myArray[index];
-  }
+    function getItemInArray(uint256 index) public view returns (uint256) {
+        return myArray[index];
+    }
 
-  function updateTheArray(uint locationInArray, uint valueToChange) public {
-    myArray[locationInArray] = valueToChange;
-  }
+    function updateTheArray(uint256 locationInArray, uint256 valueToChange) public {
+        myArray[locationInArray] = valueToChange;
+    }
 
-  function remove(uint index) public {
-    delete myArray[index];
-  }
+    function remove(uint256 index) public {
+        delete myArray[index];
+    }
 
-  function removeAndCompact(uint index) public {
-    myArray[index] = myArray[myArray.length - 1];
-    myArray.pop();
-  }
+    function removeAndCompact(uint256 index) public {
+        myArray[index] = myArray[myArray.length - 1];
+        myArray.pop();
+    }
 
-  function getLength() public view returns(uint) {
-    return myArray.length;
-  }
+    function getLength() public view returns (uint256) {
+        return myArray.length;
+    }
 }
