+++ contracts/getting_started_with_solidity/14_mapping/02_NestedMapsContract.sol
@@ -5,17 +5,21 @@
 pragma solidity ^0.8.0;
 
 contract NestedMapsContract {
-  mapping(address => mapping(address => uint)) public allowance;
+    mapping(address => mapping(address => uint256)) public allowance;
 
-  function get(address _addressOwner, address _addressSpender) public view returns(uint) {
-    return allowance[_addressOwner][_addressSpender];
-  }
+    function get(address _addressOwner, address _addressSpender) public view returns (uint256) {
+        return allowance[_addressOwner][_addressSpender];
+    }
 
-  function set(address _addressOwner, address _addressSpender, uint amount) public {
-    allowance[_addressOwner][_addressSpender] = amount;
-  }
+    function set(
+        address _addressOwner,
+        address _addressSpender,
+        uint256 amount
+    ) public {
+        allowance[_addressOwner][_addressSpender] = amount;
+    }
 
-  function remove(address _addressOwner, address _addressSpender) public {
-    delete allowance[_addressOwner][_addressSpender];
-  }
+    function remove(address _addressOwner, address _addressSpender) public {
+        delete allowance[_addressOwner][_addressSpender];
+    }
 }
