+++ contracts/getting_started_with_solidity/16_struct/MyContract.sol
@@ -5,19 +5,19 @@
 pragma solidity ^0.8.0;
 
 contract MyContract {
-  struct Todo {
-    string text;
-    bool completed;
-  }
+    struct Todo {
+        string text;
+        bool completed;
+    }
 
-  Todo[] public todos;
+    Todo[] public todos;
 
-  function create(string memory _text) public {
-    todos.push(Todo(_text, false));
-  }
+    function create(string memory _text) public {
+        todos.push(Todo(_text, false));
+    }
 
-  function get(uint _index) public view returns(string memory, bool) {
-    Todo storage todo = todos[_index];
-    return(todo.text, todo.completed);
-  }
+    function get(uint256 _index) public view returns (string memory, bool) {
+        Todo storage todo = todos[_index];
+        return (todo.text, todo.completed);
+    }
 }
