+++ contracts/getting_started_with_solidity/17_payable_modifier/MyContract.sol
@@ -5,22 +5,22 @@
 pragma solidity ^0.8.0;
 
 contract MyContract {
-  address payable public owner;
+    address payable public owner;
 
-  constructor() payable {
-    owner = payable(msg.sender);
-  }
+    constructor() payable {
+        owner = payable(msg.sender);
+    }
 
-  modifier onlyOwner() {
-    require(msg.sender == owner, "Not owner");
-    _;
-  }
+    modifier onlyOwner() {
+        require(msg.sender == owner, "Not owner");
+        _;
+    }
 
-  function withdraw(uint amount) public onlyOwner {
-    owner.transfer(amount);
-  }
+    function withdraw(uint256 amount) public onlyOwner {
+        owner.transfer(amount);
+    }
 
-  function transfer(address payable _to, uint amount) public onlyOwner {
-    _to.transfer(amount);
-  }
+    function transfer(address payable _to, uint256 amount) public onlyOwner {
+        _to.transfer(amount);
+    }
 }
