+++ contracts/getting_started_with_solidity/23_libraries/MathLibrary.sol
@@ -5,7 +5,7 @@
 pragma solidity ^0.8.0;
 
 library MathLibrary {
-  function multiply(uint a, uint b) public view returns(uint, address) {
-    return (a * b, address(this));
-  }
+    function multiply(uint256 a, uint256 b) public view returns (uint256, address) {
+        return (a * b, address(this));
+    }
 }
