+++ contracts/getting_started_with_solidity/23_libraries/MyContract.sol
@@ -9,20 +9,20 @@ import "./MathLibrary.sol";
 
 // Option 2: Library code directly in the contract
 library MathLibrary2 {
-  function multiply(uint a, uint b) internal view returns (uint, address) {
-      return (a * b, address(this));
-  }
+    function multiply(uint256 a, uint256 b) internal view returns (uint256, address) {
+        return (a * b, address(this));
+    }
 }
 
 contract MyContract {
-  using MathLibrary2 for uint;
-  address owner = address(this);
+    using MathLibrary2 for uint256;
+    address owner = address(this);
 
-  function callLibrary(uint a, uint b) public view returns(uint, address) {
-    return MathLibrary.multiply(a, b);
-  }
+    function callLibrary(uint256 a, uint256 b) public view returns (uint256, address) {
+        return MathLibrary.multiply(a, b);
+    }
 
-    function callLibrary2(uint a, uint b) public view returns(uint, address) {
-    return a.multiply(b);
-  }
+    function callLibrary2(uint256 a, uint256 b) public view returns (uint256, address) {
+        return a.multiply(b);
+    }
 }
