+++ contracts/sample_apps/01_sample_contract/MyContract.sol
@@ -5,22 +5,22 @@
 pragma solidity ^0.8.0;
 
 contract MyContract {
-  string private name;
-  uint private age;
+    string private name;
+    uint256 private age;
 
-  function setName(string memory _name) public {
-    name = _name;
-  }
+    function setName(string memory _name) public {
+        name = _name;
+    }
 
-  function getName() public view returns(string memory) {
-    return name;
-  }
+    function getName() public view returns (string memory) {
+        return name;
+    }
 
-  function setAge(uint _age) public {
-    age = _age;
-  }
+    function setAge(uint256 _age) public {
+        age = _age;
+    }
 
-  function getAge() public view returns(uint) {
-    return age;
-  }
+    function getAge() public view returns (uint256) {
+        return age;
+    }
 }
