+++ contracts/solidity_by_examples/01_ether_wallet/EtherWallet.sol
@@ -12,18 +12,18 @@ An example of a basic wallet.
 pragma solidity ^0.8.9;
 
 contract EtherWallet {
-  address payable public owner;
+    address payable public owner;
 
-  constructor() payable {
-    owner = payable(msg.sender);
-  }
+    constructor() payable {
+        owner = payable(msg.sender);
+    }
 
-  function withdraw(uint amount) external {
-    require(msg.sender == owner, "caller is not owner");
-    payable(owner).transfer(amount);
-  }
+    function withdraw(uint256 amount) external {
+        require(msg.sender == owner, "caller is not owner");
+        payable(owner).transfer(amount);
+    }
 
-  function getBalance() external view returns(uint) {
-    return address(this).balance;
-  }
+    function getBalance() external view returns (uint256) {
+        return address(this).balance;
+    }
 }
